Refactor the PhatPages middleware
Before, PhatPages always used the same built in view, but allowed users
to define a custom template to use with that view. Because we need to 
pass custom analytics variables, etc, we need the ability to have custom
view instead of a custom template.

Custom views are defined in the `settings.PHATPAGE_VIEW` variable and
may be the string path to either a view function or view class.